[#2] refactor: components/filter폴더 추가
components/filter폴더내에 파일들이 많아져 pricefilter, spacefilter 폴더를 만들어 관련된것들을 이동 시켜줬습니다.
